requirements.txt
Flask
Flask-Cors
streamlit
ultralytics
numpy
Pillow
opencv-python-headless
pyopenssl